Moved creator functions to utils
