+++ client/src/app/utils/HomeSettings.ts
